refactor: moved hooks to a more specific directory
