+++ api/event/getCategories.ts
