+++ hooks/event/useInterests.ts
@@ -1,7 +1,7 @@
 import {useQuery} from "@tanstack/react-query";
 import {Interest} from "@/types/types";
 import {GET_INTERESTS} from "@/constants/queryKey";
-import getInterests from "@/api/getInterests";
+import getInterests from "@/api/event/getInterests";
 
 
 const useInterests = () => {
